+++ orix/plot/stereographic_plot.py
@@ -27,13 +27,13 @@ from matplotlib.transforms import Affine2D, Affine2DBase, BboxTransformTo, Trans
 import numpy as np
 
 from orix.projections import InverseStereographicProjection, StereographicProjection
-from orix.vector import Vector3d
 from orix.plot._symmetry_marker import (
     TwoFoldMarker,
     ThreeFoldMarker,
     FourFoldMarker,
     SixFoldMarker,
 )
+from orix.vector import Vector3d
 
 
 class StereographicTransform(Transform):
@@ -97,13 +97,16 @@ class StereographicAffine(Affine2DBase):
         if self._invalid:
             pole = self.pole
             st = StereographicTransform(pole=pole)
-            xscale, _ = st.transform((0, 0.5 * np.pi))
-            _, yscale = st.transform((0.5 * np.pi, 0.5 * np.pi))
+            xscale, _ = st.transform((0, np.pi / 2))
+            _, yscale = st.transform((np.pi / 2, np.pi / 2))
             scales = (0.5 / xscale, 0.5 / yscale)
             self._mtx = Affine2D().scale(*scales).translate(0.5, 0.5)
             self._inverted = None
             self._invalid = 0
         return self._mtx
+
+
+ZORDER = dict(text=6, scatter=5, symmetry_marker=4, draw_circle=3)
 
 
 class StereographicPlot(Axes):
@@ -130,12 +133,12 @@ class StereographicPlot(Axes):
     name = "stereographic"
     _hemisphere = "upper"
 
-    def __init__(self, *args, polar_resolution=15, azimuth_resolution=15, **kwargs):
-        self._polar_cap = 0.5 * np.pi
-        self._polar_resolution = polar_resolution
-
+    def __init__(self, *args, azimuth_resolution=10, polar_resolution=10, **kwargs):
         self._azimuth_cap = 2 * np.pi
         self._azimuth_resolution = azimuth_resolution
+
+        self._polar_cap = np.pi / 2
+        self._polar_resolution = polar_resolution
 
         super().__init__(*args, **kwargs)
         # Set ratio of y-unit to x-unit by adjusting the physical
@@ -225,7 +228,7 @@ class StereographicPlot(Axes):
 
         Parameters
         ----------
-        args : orix.vector.Vector3d or tuple of float or numpy.ndarray
+        args : Vector3d or tuple of float or numpy.ndarray
             Vector(s), or azimuth and polar angles, the latter two
             passed as separate arguments (not keyword arguments).
         kwargs
@@ -236,11 +239,26 @@ class StereographicPlot(Axes):
         --------
         matplotlib.axes.Axes.scatter
         """
-        new_kwargs = dict(zorder=3, clip_on=False)
+        new_kwargs = dict(zorder=ZORDER["scatter"], clip_on=False)
         for k, v in new_kwargs.items():
             kwargs.setdefault(k, v)
+
         azimuth, polar = self._pretransform_input(args)
-        super().scatter(azimuth, polar, **kwargs)
+        visible = self._visible_in_hemisphere(polar)
+        if np.count_nonzero(visible) == 0:
+            return
+        else:
+            azimuth = azimuth[visible]
+            polar = polar[visible]
+
+        # Color(s) and size(s)
+        c = kwargs.pop("c", "C0")
+        c = _get_array_of_values(value=c, visible=visible)
+        s = kwargs.pop("s", None)
+        if s is not None:
+            s = _get_array_of_values(value=s, visible=visible)
+
+        super().scatter(azimuth, polar, c=c, s=s, **kwargs)
 
     def text(self, *args, **kwargs):
         """Add text to the axes.
@@ -250,7 +268,7 @@ class StereographicPlot(Axes):
 
         Parameters
         ----------
-        args : orix.vector.Vector3d or tuple of float or numpy.ndarray
+        args : Vector3d or tuple of float or numpy.ndarray
             Vector(s), or azimuth and polar angles, the latter two
             passed as separate arguments (not keyword arguments).
         kwargs
@@ -261,10 +279,18 @@ class StereographicPlot(Axes):
         --------
         matplotlib.axes.Axes.text
         """
-        new_kwargs = dict(va="bottom", ha="center")
+        new_kwargs = dict(va="bottom", ha="center", zorder=ZORDER["text"])
         for k, v in new_kwargs.items():
             kwargs.setdefault(k, v)
+
         azimuth, polar = self._pretransform_input(args)
+        visible = self._visible_in_hemisphere(polar)
+        if np.count_nonzero(visible) == 0:
+            return
+        else:
+            azimuth = azimuth[visible]
+            polar = polar[visible]
+
         super().text(azimuth, polar, **kwargs)
 
     # ----------- Custom attributes and methods below here ----------- #
@@ -318,7 +344,7 @@ class StereographicPlot(Axes):
         """
         new_kwargs = dict(ha="right", va="bottom")
         new_kwargs.update(kwargs)
-        Axes.text(self, 0.75 * np.pi, 0.5 * np.pi, s=self.hemisphere, **new_kwargs)
+        Axes.text(self, 0.75 * np.pi, np.pi / 2, s=self.hemisphere, **new_kwargs)
 
     def azimuth_grid(self, resolution=None):
         """Set the azimuth grid resolution in degrees.
@@ -366,19 +392,19 @@ class StereographicPlot(Axes):
         new_kwargs.update(kwargs)
         super().text(x=x, y=y, s=label, **new_kwargs)
 
-    def set_labels(self, xlabel="X", ylabel="Y", zlabel="Z", **kwargs):
+    def set_labels(self, xlabel="x", ylabel="y", zlabel="z", **kwargs):
         """Set the reference frame's axes labels.
 
         Parameters
         ----------
         xlabel : str, False or None, optional
-            X axis label, default is "X". If False or None, this label
+            X axis label, default is "x". If False or None, this label
             is not shown.
         ylabel : str, False or None, optional
-            Y axis label, default is "Y". If False or None, this label
+            Y axis label, default is "y". If False or None, this label
             is not shown.
         zlabel : str, False or None, optional
-            Z axis label, default is "Z". If False or None, this label
+            Z axis label, default is "z". If False or None, this label
             is not shown.
         """
         # z label position
@@ -403,17 +429,20 @@ class StereographicPlot(Axes):
         kwargs
             Keyword arguments passed to :meth:`scatter`.
         """
-        size = kwargs.pop("s", 1)
-        if fold == 2:
-            marker = TwoFoldMarker(v, size=size)
-        elif fold == 3:
-            marker = ThreeFoldMarker(v, size=size)
-        elif fold == 4:
-            marker = FourFoldMarker(v, size=size)
-        elif fold == 6:
-            marker = SixFoldMarker(v, size=size)
-        else:
+        if fold not in [2, 3, 4, 6]:
             raise ValueError("Can only plot 2-, 3-, 4- or 6-fold elements.")
+
+        marker_classes = {
+            "2": TwoFoldMarker,
+            "3": ThreeFoldMarker,
+            "4": FourFoldMarker,
+            "6": SixFoldMarker,
+        }
+        marker = marker_classes[str(fold)](v, size=kwargs.pop("s", 1))
+
+        new_kwargs = dict(zorder=ZORDER["symmetry_marker"], clip_on=False)
+        for k, v in new_kwargs.items():
+            kwargs.setdefault(k, v)
 
         for vec, marker, marker_size in marker:
             self.scatter(vec, marker=marker, s=marker_size, **kwargs)
@@ -421,23 +450,196 @@ class StereographicPlot(Axes):
         # TODO: Find a way to control padding, so that markers aren't
         #  clipped
 
+    def draw_circle(self, *args, opening_angle=np.pi / 2, steps=100, **kwargs):
+        r"""Draw great or small circles with a given `opening_angle` to
+        one or multiple vectors.
+
+        A vector must be present in the current hemisphere for its
+        circle to be drawn.
+
+        Parameters
+        ----------
+        args : Vector3d or tuple of float or numpy.ndarray
+            Vector(s), or azimuth and polar angles defining vectors, the
+            latter two passed as separate arguments (not keyword
+            arguments). Circles are drawn perpendicular to these with a
+            given `opening_angle`.
+        opening_angle : float or numpy.ndarray, optional
+            Opening angle(s) around the vector(s). Default is
+            :math:`\pi/2`, giving a great circle. If an array is passed,
+            its size must be equal to the number of circles to draw.
+        steps : int, optional
+            Number of vectors to describe each circle, default is 100.
+        kwargs
+            Keyword arguments passed to
+            :meth:`matplotlib.axes.Axes.plot` to alter the circles'
+            appearance.
+
+        See Also
+        --------
+        orix.vector.Vector3d.get_circle
+        """
+        azimuth, polar = self._pretransform_input(args)
+
+        # Exclude vectors not visible in this hemisphere before creating
+        # circles
+        visible = self._visible_in_hemisphere(polar)
+        if np.count_nonzero(visible) == 0:  # No circles to draw
+            return
+        else:
+            azimuth = azimuth[visible]
+            polar = polar[visible]
+
+        # Get set of `steps` vectors delineating a circle per vector
+        v = Vector3d.from_polar(azimuth=azimuth, polar=polar)
+        circles = v.get_circle(opening_angle=opening_angle, steps=steps).unit
+
+        # Enable using one color per circle
+        color = kwargs.pop("color", "C0")
+        color2 = _get_array_of_values(value=color, visible=visible)
+
+        # Set above which elements circles will appear (zorder)
+        new_kwargs = dict(zorder=ZORDER["draw_circle"], clip_on=False)
+        for k, v in new_kwargs.items():
+            kwargs.setdefault(k, v)
+
+        hemisphere = self.hemisphere
+        polar_cap = self._polar_cap
+        for i, c in enumerate(circles):
+            a, p = _sort_coords_by_shifted_bools(
+                hemisphere=hemisphere,
+                polar_cap=polar_cap,
+                azimuth=c.azimuth.data,
+                polar=c.polar.data,
+            )
+            super().plot(a, p, color=color2[i], **kwargs)
+
     @staticmethod
     def _pretransform_input(values):
-        """Return azimuth and polar angles from input data."""
+        """Return arrays of azimuth and polar angles from input data.
+
+        Parameters
+        ----------
+        values : tuple of numpy.ndarray or Vector3d
+            Spherical coordinates (azimuth, polar) or vectors. If
+            spherical coordinates are given, they are assumed to
+            describe unit vectors.
+
+        Returns
+        -------
+        azimuth : numpy.ndarray
+            Azimuth coordiantes of unit vectors.
+        polar : numpy.ndarray
+            Polar coordinates of unit vectors.
+        """
         if len(values) == 2:
-            azimuth, polar = values
+            azimuth = np.asarray(values[0])
+            polar = np.asarray(values[1])
         else:
-            value = values[0]
-            if isinstance(value, Vector3d):
-                value = value.unit
+            try:
+                value = values[0].unit
                 azimuth = value.azimuth.data
                 polar = value.polar.data
-            else:
+            except (ValueError, AttributeError):
                 raise ValueError(
                     "Accepts only one (Vector3d) or two (azimuth, polar) input "
                     "arguments"
                 )
         return azimuth, polar
 
+    def _visible_in_hemisphere(self, polar):
+        """Return a boolean array describing whether the vector which
+        the polar angles belong to are visible in the current
+        hemisphere.
+
+        Parameters
+        ----------
+        polar : numpy.ndarray
+
+        Returns
+        -------
+        numpy.ndarray
+            Boolean array with True for the polar angles corresponding
+            to vectors visible in this hemisphere.
+        """
+        return _visible_in_hemisphere(self.hemisphere, self._polar_cap, polar)
+
 
 register_projection(StereographicPlot)
+
+
+def _get_array_of_values(value, visible):
+    """Return a usable array of `value` with the correct size
+    even though `value` doesn't have as many elements as `visible.size`,
+    to be iterated over along with `True` elements in `visible`.
+
+    Parameters
+    ----------
+    value : str, float, or a list of str or float
+        Typically a keyword argument value to be passed to some
+        Matplotlib routine.
+    visible : numpy.ndarray
+        Boolean array with as many elements as input vectors, only some
+        of which are visible in the hemisphere (`True`).
+
+    Returns
+    -------
+    numpy.ndarray
+        An array populated with `value` of a size equal to the number of
+        True elements in `visible`.
+    """
+    n = visible.size
+    if not isinstance(value, str) and hasattr(value, "__iter__") and len(value) != n:
+        value = value[0]
+    if isinstance(value, str) or not hasattr(value, "__iter__"):
+        value = [value,] * n
+    return np.asarray(value)[visible]
+
+
+def _visible_in_hemisphere(hemisphere, polar_cap, polar):
+    """Return a boolean array describing whether the vector which the
+    polar angles belong to are visible in the current hemisphere.
+
+    Parameters
+    ----------
+    hemisphere : str
+    polar_cap : float
+    polar : numpy.ndarray
+
+    Returns
+    -------
+    numpy.ndarray
+        Boolean array with True for polar angles corresponding to
+        vectors visible in this hemisphere.
+    """
+    return polar <= polar_cap if hemisphere == "upper" else polar > polar_cap
+
+
+def _sort_coords_by_shifted_bools(hemisphere, polar_cap, azimuth, polar):
+    """Shift azimuth and polar coordinate arrays so that the ones
+    corresponding to vectors visible in this hemisphere are shifted to
+    the start of the arrays.
+
+    Used in :meth:`StereographicPlot.draw_circle`.
+
+    Parameters
+    ----------
+    hemisphere : str
+    polar_cap : float
+    azimuth : numpy.ndarray
+    polar : numpy.ndarray
+
+    Returns
+    -------
+    azimuth : numpy.ndarray
+    polar : numpy.ndarray
+    """
+    visible = _visible_in_hemisphere(
+        hemisphere=hemisphere, polar_cap=polar_cap, polar=polar
+    )
+    indices = np.asarray(visible != visible[0]).nonzero()[0]
+    if indices.size != 0:
+        to_shift = indices[-1] + 1
+        azimuth = np.roll(azimuth, shift=-to_shift)
+        polar = np.roll(polar, shift=-to_shift)
+    return azimuth, polar